refactor(editor): extract property edit planning
Move Route Net Label lookup and naming, Power Label naming, external-device parameter discovery, and instance property transaction planning out of App. The existing Properties hook retains draft ownership and the transaction controller remains the mutation boundary.

Validation: pnpm ci:static; pnpm test:local apps/editor/src/features/properties/property-edit-planner.test.ts apps/editor/src/features/properties/use-properties-editor.test.tsx apps/editor/src/app/App.test.tsx; pnpm test:e2e:local apps/editor/e2e/manual-editor.spec.ts --grep relevant Net-label/property/value scenarios (4 passed).

Test-Impact: tests-updated
