+++ apps/editor/src/app/App.tsx
@@ -23,7 +23,6 @@ import {
   proposePowerRailTranslation,
   proposeWireCommitThroughContacts,
   proposeWireSegmentMove,
-  planEnsureNamedNet,
   planDeleteCell,
   planRemoveCellTerminalMarkers,
   planEditCellTerminalAnnotation,
@@ -58,7 +57,6 @@ import {
   resolveDraftingObjectGeometry,
   resolveElectricalContactTargets,
   displayableInstanceValue,
-  resolveNetLabelBinding,
   resolveMosBulkConnection,
   resolveDocumentStyleProfile,
   resolveDocumentLogicalNets,
@@ -80,7 +78,6 @@ import type {
 import {
   createEmptyProject,
   createId,
-  deriveStableId,
   defaultDraftTextDocument,
   foldNetName,
   flattenRichText,
@@ -92,13 +89,11 @@ import {
 import type {
   Annotation,
   CircuitProject,
-  ConnectivityEvidence,
   DerivedPoint,
   DraftingObject,
   GridRect,
   Point,
   Rect,
-  RichTextDocument,
   RouteEndpoint,
   SchematicDocument,
 } from "@icm/model";
@@ -167,10 +162,6 @@ import {
   proposedStandalonePowerConnection,
 } from "../features/component-insert/placement-connectivity";
 import {
-  componentParameters,
-  externalMosComponentParameters,
-} from "../features/component-insert/component-parameters";
-import {
   endpointTestId,
   instanceLabelAnnotationFor,
   maxRoutingCounter,
@@ -198,7 +189,6 @@ import {
 } from "./lazy-editor-dialogs";
 import {
   bindingForEditedModel,
-  initialInstanceNetlist,
   netlistReferenceMatchesPlacement,
   nextInstanceDesignator,
 } from "../features/netlist-export/netlist-authoring";
@@ -316,6 +306,7 @@ import {
   type SupplementalSelection,
 } from "../features/selection/selection-inspection-model";
 import { usePropertiesEditor } from "../features/properties/use-properties-editor";
+import { createPropertyEditPlanner } from "../features/properties/property-edit-planner";
 import {
   LIBRARY_WIDTH_MAX,
   LIBRARY_WIDTH_MIN,
@@ -1085,6 +1076,17 @@ export function App({
     () => buildProjectConnectivityIndex(project, resolver),
     [project, resolver],
   );
+  const routeGeometryRecords = useMemo(
+    () =>
+      document.routes.flatMap((route) => {
+        const geometry = projectConnectivityIndex.documents
+          .get(document.id)
+          ?.routingGeometry.routes.get(route.id);
+        if (!geometry) return [];
+        return [{ route, geometry }];
+      }),
+    [document, projectConnectivityIndex],
+  );
   const netlistAnalysis = useMemo(
     () => analyzeDesignNetlist(project),
     [project],
@@ -1174,6 +1176,19 @@ export function App({
     selectedHierarchyCell,
     selectedInstance,
   ]);
+  const {
+    netLabelForRoute,
+    netLabelEditsForRoute,
+    netNameEditsForAnnotation,
+    propertyParametersForInstance,
+    instancePropertyEdits,
+  } = createPropertyEditPlanner({
+    project,
+    document,
+    resolver,
+    routeGeometryRecords,
+    setStatus,
+  });
   const {
     addAdditionalParameter,
     additionalParameterDraft,
@@ -1520,17 +1535,6 @@ export function App({
     }
     return [...byKey.values()];
   }, [visibleBulkEndpoints, visibleEndpoints]);
-  const routeGeometryRecords = useMemo(
-    () =>
-      document.routes.flatMap((route) => {
-        const geometry = projectConnectivityIndex.documents
-          .get(document.id)
-          ?.routingGeometry.routes.get(route.id);
-        if (!geometry) return [];
-        return [{ route, geometry }];
-      }),
-    [document, projectConnectivityIndex],
-  );
   const contactComponents = useMemo(
     () =>
       [
@@ -4874,211 +4878,6 @@ export function App({
     }
   }
 
-  function netLabelForRoute(
-    route: SchematicDocument["routes"][number],
-  ): Annotation | undefined {
-    const candidates = document.annotations.filter(
-      (annotation) =>
-        annotation.kind === "net-label" && annotation.netId === route.netId,
-    );
-    return (
-      candidates.find(
-        (annotation) => annotation.id === `net-label-${route.id}`,
-      ) ??
-      candidates.find(
-        (annotation) =>
-          resolveNetLabelBinding(document, resolver, annotation)?.routeId ===
-          route.id,
-      )
-    );
-  }
-
-  function netLabelEditsForRoute(
-    route: SchematicDocument["routes"][number],
-    rawName: string,
-    presentation?: {
-      alignment: "start" | "middle" | "end";
-      sizeScale: number;
-      formatOverride?: RichTextDocument;
-    },
-  ): SchematicEdit[] | null {
-    const net = document.nets.find((candidate) => candidate.id === route.netId);
-    if (!net) return null;
-    const existingLabel = netLabelForRoute(route);
-    const name = rawName.trim();
-    if (!name) {
-      return existingLabel
-        ? [
-            {
-              kind: "remove_schematic_annotation",
-              annotationId: existingLabel.id,
-            },
-          ]
-        : null;
-    }
-    const labelId = existingLabel?.id ?? `net-label-${route.id}`;
-    const namedNetPlan = planEnsureNamedNet(document, {
-      candidateNetId: net.id,
-      name,
-      evidenceId:
-        document.connectivityEvidence.find(
-          (evidence) =>
-            evidence.kind === "name-claim" &&
-            evidence.owner.kind === "net-label" &&
-            evidence.owner.annotationId === labelId,
-        )?.id ??
-        deriveStableId(
-          "connectivity-evidence",
-          document.id,
-          "net-label",
-          net.id,
-          labelId,
-        ),
-      owner: { kind: "net-label", annotationId: labelId },
-    });
-    if (!namedNetPlan.ok) return null;
-    const targetNetId = namedNetPlan.netId;
-    const geometry = routeGeometryRecords.find(
-      ({ route: candidate }) => candidate.id === route.id,
-    )?.geometry;
-    if (!geometry) return null;
-    const segment = Math.max(
-      0,
-      Math.floor((geometry.centerline.length - 1) / 2),
-    );
-    const from = geometry.centerline[segment]!;
-    const to = geometry.centerline[segment + 1] ?? from;
-    const position = snapGridPoint(
-      (existingLabel
-        ? existingLabel.anchor.kind === "free"
-          ? existingLabel.anchor.position
-          : existingLabel.anchor.fallbackPosition
-        : undefined) ?? {
-        x: (from.x + to.x) / 2,
-        y: (from.y + to.y) / 2 - 8,
-      },
-      document.presentation.grid,
-    );
-    const previousAnchor =
-      existingLabel?.anchor.kind === "route" &&
-      existingLabel.anchor.routeId === route.id
-        ? existingLabel.anchor
-        : null;
-    const edits: SchematicEdit[] = [...namedNetPlan.edits];
-    edits.push({
-      kind: "upsert_schematic_annotation",
-      annotation: {
-        id: labelId,
-        kind: "net-label",
-        binding: { kind: "net-name", netId: targetNetId },
-        netId: targetNetId,
-        // A dragged route anchor survives a name edit; new labels start at
-        // the middle segment with the default normal offset.
-        anchor: previousAnchor
-          ? { ...previousAnchor, fallbackPosition: position }
-          : {
-              kind: "route",
-              routeId: route.id,
-              segmentIndex: segment,
-              t: 0.5,
-              normalOffset: -8,
-              direction: "forward",
-              orientation: "follow",
-              fallbackPosition: position,
-            },
-        alignment:
-          presentation?.alignment ?? existingLabel?.alignment ?? "middle",
-        rotation: 0,
-        locked: false,
-        ...(presentation?.sizeScale !== undefined
-          ? { sizeScale: presentation.sizeScale }
-          : existingLabel?.sizeScale !== undefined
-            ? { sizeScale: existingLabel.sizeScale }
-            : {}),
-        ...(presentation?.formatOverride
-          ? { formatOverride: presentation.formatOverride }
-          : {}),
-      },
-    });
-    return edits;
-  }
-
-  function netNameEditsForAnnotation(
-    annotation: Annotation,
-    rawName: string,
-    presentationAnnotation?: Annotation,
-  ): SchematicEdit[] | null | undefined {
-    const binding = annotation.binding;
-    if (binding?.kind !== "net-name" || annotation.anchor.kind !== "object") {
-      return undefined;
-    }
-    const ownerObjectId = annotation.anchor.objectId;
-    const instance = document.instances.find(
-      (candidate) => candidate.id === ownerObjectId,
-    );
-    const powerClaim = document.connectivityEvidence.find(
-      (
-        evidence,
-      ): evidence is Extract<ConnectivityEvidence, { kind: "name-claim" }> =>
-        evidence.kind === "name-claim" &&
-        evidence.owner.kind === "power-marker" &&
-        (evidence.owner.objectId === ownerObjectId ||
-          evidence.owner.objectId === annotation.id),
-    );
-    if (annotation.kind !== "power-label") return undefined;
-    const net = document.nets.find(
-      (candidate) => candidate.id === binding.netId,
-    );
-    if (!net) {
-      setStatus(`Net Label references missing Net ${binding.netId}`);
-      return null;
-    }
-    const name = rawName.trim();
-    const namedNetPlan = planEnsureNamedNet(document, {
-      candidateNetId: net.id,
-      name,
-      evidenceId:
-        powerClaim?.id ??
-        deriveStableId(
-          "connectivity-evidence",
-          document.id,
-          "power-marker",
-          net.id,
-          annotation.id,
-        ),
-      owner: {
-        kind: "power-marker",
-        objectId:
-          powerClaim?.owner.kind === "power-marker"
-            ? powerClaim.owner.objectId
-            : annotation.id,
-      },
-      ...(powerClaim
-        ? {
-            scope: powerClaim.scope,
-            ...(powerClaim.powerDomain
-              ? { powerDomain: powerClaim.powerDomain }
-              : {}),
-          }
-        : {}),
-    });
-    if (!namedNetPlan.ok) {
-      setStatus(namedNetPlan.message);
-      return null;
-    }
-    return [
-      ...namedNetPlan.edits,
-      ...(presentationAnnotation
-        ? [
-            {
-              kind: "upsert_schematic_annotation" as const,
-              annotation: presentationAnnotation,
-            },
-          ]
-        : []),
-    ];
-  }
-
   function referenceLabelVisibilityEdits(
     instanceIds: readonly string[],
     visible: boolean,
@@ -5155,92 +4954,6 @@ export function App({
     return edits;
   }
 
-  function instancePropertyEdits(draft: {
-    instanceId: string | null;
-    parameters: Record<string, string>;
-    x: string;
-    y: string;
-    rotation: "0" | "90" | "180" | "270";
-  }): { edits: SchematicEdit[]; invalidPosition: boolean } {
-    if (!draft.instanceId) return { edits: [], invalidPosition: false };
-    const instance = document.instances.find(
-      (item) => item.id === draft.instanceId,
-    );
-    if (!instance) return { edits: [], invalidPosition: false };
-    const edits: SchematicEdit[] = [];
-    const baseNetlist =
-      instance.netlist ??
-      initialInstanceNetlist(document, instance.symbolId, {});
-    if (baseNetlist) {
-      const netlistParameters = { ...baseNetlist.parameters };
-      const set: Record<string, string> = {};
-      const unset: string[] = [];
-      for (const parameter of propertyParametersForInstance(instance)) {
-        const value = (draft.parameters[parameter.key] ?? "").trim();
-        const current = netlistParameters[parameter.key];
-        if (value === "") {
-          delete netlistParameters[parameter.key];
-          if (current !== undefined) unset.push(parameter.key);
-        } else {
-          netlistParameters[parameter.key] = value;
-          if (current !== value) set[parameter.key] = value;
-        }
-      }
-
-      const nextNetlist = {
-        ...baseNetlist,
-        parameters: netlistParameters,
-      };
-      if (!instance.netlist) {
-        edits.push({
-          kind: "set_instance_netlist",
-          instanceId: instance.id,
-          netlist: nextNetlist,
-        });
-      } else if (Object.keys(set).length > 0 || unset.length > 0) {
-        edits.push({
-          kind: "patch_instance_netlist_parameters",
-          instanceId: instance.id,
-          ...(Object.keys(set).length > 0 ? { set } : {}),
-          ...(unset.length > 0 ? { unset } : {}),
-        });
-      }
-    }
-
-    let invalidPosition = false;
-    if (instance.placement) {
-      const x = Number(draft.x);
-      const y = Number(draft.y);
-      if (!Number.isFinite(x) || !Number.isFinite(y)) {
-        invalidPosition = true;
-      } else {
-        const position = {
-          x: snapCoordinate(x, document.presentation.grid),
-          y: snapCoordinate(y, document.presentation.grid),
-        };
-        if (
-          position.x !== instance.placement.position.x ||
-          position.y !== instance.placement.position.y
-        ) {
-          edits.push({
-            kind: "move_instance",
-            instanceId: instance.id,
-            position,
-          });
-        }
-      }
-      const rotation = Number(draft.rotation) as 0 | 90 | 180 | 270;
-      if (rotation !== instance.placement.rotation) {
-        edits.push({
-          kind: "rotate_instance",
-          instanceId: instance.id,
-          rotation,
-        });
-      }
-    }
-    return { edits, invalidPosition };
-  }
-
   function updateSelectedModelTarget(value: string): void {
     if (!selectedInstance?.netlist) return;
     if (
@@ -5296,32 +5009,6 @@ export function App({
           : `Cleared model target for ${selectedInstance.id}`,
       );
     }
-  }
-
-  function propertyParametersForInstance(
-    instance: SchematicDocument["instances"][number],
-  ) {
-    const binding = instance.netlist?.binding;
-    if (binding?.kind === "external-subcircuit") {
-      const definition = project.externalSubcircuitDefinitions.find(
-        (candidate) => candidate.id === binding.definitionId,
-      );
-      const mapping = definition
-        ? resolvePdkSymbolMapping(definition.name, definition.terminals.length)
-        : undefined;
-      if (
-        definition &&
-        (mapping?.symbolId === "nmos" || mapping?.symbolId === "pmos") &&
-        definition.terminals.every(
-          (terminal, index) =>
-            terminal.name.toLowerCase() ===
-            mapping.pinNames[index]?.toLowerCase(),
-        )
-      ) {
-        return externalMosComponentParameters(mapping.symbolId);
-      }
-    }
-    return componentParameters(instance.symbolId);
   }
 
   function updateSelectedSchematicName(value: string): void {
